Fix bitmap pooling test
